Refactor to abstract test, (re)set error reporting

+++ test/classes/phing/tasks/ext/svn/AbstractSvnTaskTest.php
@@ -20,24 +20,37 @@
  */
  
 require_once 'phing/BuildFileTest.php';
-require_once 'phing/tasks/ext/svn/AbstractSvnTaskTest.php';
+require_once dirname(__FILE__) . '/../GitTasks/GitTestsHelper.php';
 
 /**
  * @author Michiel Rook <mrook@php.net>
  * @version $Id$
  * @package phing.tasks.ext
  */
-class SvnSwitchTaskTest extends AbstractSvnTaskTest { 
-    public function setUp() {
-        parent::setUp('SvnSwitchTest.xml');
-        GitTestsHelper::rmdir(PHING_TEST_BASE . '/tmp/svn');
+abstract class AbstractSvnTaskTest extends BuildFileTest { 
+    protected $backupGlobals = FALSE;
+    
+    protected $savedErrorLevel = 0;
+     
+    public function setUp($buildFilename) { 
+        if (is_readable(PHING_TEST_BASE . '/tmp/svn')) {
+            // make sure we purge previously created directory
+            // if left-overs from previous run are found
+            GitTestsHelper::rmdir(PHING_TEST_BASE . '/tmp/svn');
+        }
+        // set temp directory used by test cases
+        mkdir(PHING_TEST_BASE . '/tmp/svn');
+
+        $this->savedErrorLevel = error_reporting();
+        error_reporting(E_ERROR);
+        
+        $this->configureProject(PHING_TEST_BASE 
+                              . '/etc/tasks/ext/svn/' . $buildFilename);
     }
 
-    public function testSwitchSimple()
+    public function tearDown()
     {
-        $repository = PHING_TEST_BASE . '/tmp/svn';
-        $this->executeTarget('switchSimple');
-        $this->assertInLogs("Checking out SVN repository to '" . $repository . "'");
-        $this->assertInLogs("Switching SVN repository at '$repository' to 'https://github.com/phingofficial/phing/tags/2.4.12/etc'");
+        error_reporting($this->savedErrorLevel);
+        GitTestsHelper::rmdir(PHING_TEST_BASE . '/tmp/svn');
     }
 }
